refactored out DataFunctions
